--- nuxt-pinia-tailwind/types/auth/interfaces.ts
@@ -1,59 +1,0 @@
-export interface IUserOrg {
-  login: string
-  id: number
-  node_id: string
-  url: string
-  repos_url: string
-  events_url: string
-  hooks_url: string
-  issues_url: string
-  members_url: string
-  public_members_url: string
-  avatar_url: string
-  description: string
-}
-
-export interface IUserStar {
-  name: string
-  url: string
-}
-
-export interface IUserApiResponse {
-  login: string
-  id: number
-  node_id: string
-  avatar_url: string
-  gravatar_id: string
-  url: string
-  html_url: string
-  followers_url: string
-  following_url: string
-  gists_url: string
-  starred_url: string
-  subscriptions_url: string
-  organizations_url: string
-  repos_url: string
-  events_url: string
-  received_events_url: string
-  type: string
-  site_admin: boolean
-  name: string
-  company: string
-  blog: string
-  location: string
-  email: string
-  hireable: boolean
-  bio: string
-  twitter_username: string
-  public_repos: number
-  public_gists: number
-  followers: number
-  following: number
-  created_at: string
-  updated_at: string
-}
-
-export interface IUser extends IUserApiResponse {
-  orgs: IUserOrg[]
-  stars: IUserStar[]
-}
