+++ nuxt-pinia-tailwind/types/user/interfaces.ts
@@ -1,25 +1,3 @@
-export interface IUserRepository {
-  id: number
-  name: string
-  full_name: string
-  owner: { login: string }
-  description: string
-  private: boolean
-  html_url: string
-  url: string
-  updated_at: Date
-  stargazers_count: number
-  language: string
-  branches_url: string
-  visibility: 'public' | 'private'
-  subscribers_count: number
-  forks_count: number
-  open_issues_count: number
-  pulls: number
-  default_branch: string
-  homepage: string
-}
-
 export interface UserGist {
   url: string
   id: string
@@ -29,4 +7,64 @@ export interface UserGist {
       filename: string
     }
   }
+}
+
+export interface IUserOrg {
+  login: string
+  id: number
+  node_id: string
+  url: string
+  repos_url: string
+  events_url: string
+  hooks_url: string
+  issues_url: string
+  members_url: string
+  public_members_url: string
+  avatar_url: string
+  description: string
+}
+
+export interface IUserStar {
+  name: string
+  url: string
+}
+
+export interface IUserApiResponse {
+  login: string
+  id: number
+  node_id: string
+  avatar_url: string
+  gravatar_id: string
+  url: string
+  html_url: string
+  followers_url: string
+  following_url: string
+  gists_url: string
+  starred_url: string
+  subscriptions_url: string
+  organizations_url: string
+  repos_url: string
+  events_url: string
+  received_events_url: string
+  type: string
+  site_admin: boolean
+  name: string
+  company: string
+  blog: string
+  location: string
+  email: string
+  hireable: boolean
+  bio: string
+  twitter_username: string
+  public_repos: number
+  public_gists: number
+  followers: number
+  following: number
+  created_at: string
+  updated_at: string
+}
+
+export interface IUser extends IUserApiResponse {
+  orgs: IUserOrg[]
+  stars: IUserStar[]
 }
